\name{MetNorm-package}
\alias{MetNorm-package}
\alias{MetNorm}
\docType{package}
\title{ MetNorm}
\description{Statistical methods for normalizing metabolomics data }

\details{Metabolomics data are inevitably subject to a component of unwanted variation, due to factors such as 
batch effects, matrix effects, and confounding biological variation. This package contains a collection of R 
fuctions which may be used to obtain normalized metabolomics data.}

\author{Alysha M De Livera}

\references{
De Livera, A. M., Dias, D. A, De Souza, D., Rupasinghe, T., Pyke, J., Tull, D., Roessner, U., McConville, M., and Speed, T. P. (2012). Normalizing and integrating metabolomics data. Analytical chemistry, 84(24), 10768-76.

De Livera, A.M., Aho-Sysi, M., Jacob, L., Gagnon-Bartch, J., Castillo, S., Simpson, J.A., and Speed, T.P. (2014), Statistical methods for handling unwanted variation in metabolomics data
}


