\name{normdata-class}
\alias{normdata-class}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Normalized data -class}
\description{A list containing the following components:}
\section{Slots/Components}{
  \describe{
      \item{\code{unadjY}}{The unadjusted data matrix}
      
      \item{\code{newY}}{The normalized data matrix).}
      
      \item{\code{UVcomp}}{The removed unwanted variation component}
      
      \item{\code{ctl}}{A logical vector indicating the controls used in the model}
      \item{\code{k}}{The number of factors of unwanted variation}
      \item{\code{lambda}}{The regularization parameter}
      
    }

}

\author{ Alysha M De Livera}


%% ~Make other sections like Warning with \section{Warning }{....} ~

